Add chapter 1-4
All text has been typogrified, sementicated and checked with the source
material.

This chapter includes a table, and thus includes a couple rules to make
it look good. All cells are distributed to the width of it's container,
numbers are aligned left and letters are aligned right. I would prefer
to do this using <col>/<colgroup>, but the HTML standard only allows
certain styling attributes for these tags.